fix: swagger conflict #70
