--- backend/src/main/java/com/fiveguys/fivelogbackend/global/config/SwaggerConfig.java
@@ -1,25 +1,0 @@
-package com.fiveguys.fivelogbackend.global.config;
-
-import io.swagger.v3.oas.annotations.OpenAPIDefinition;
-import io.swagger.v3.oas.annotations.enums.SecuritySchemeType;
-import io.swagger.v3.oas.annotations.info.Info;
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
-import io.swagger.v3.oas.annotations.security.SecurityScheme;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@OpenAPIDefinition(
-        info = @Info(title = "fivelogs API",
-                version = "1.0",
-                description = "fiveguys 를 위한 API 문서"),
-        security = @SecurityRequirement(name = "bearerAuth")
-)
-@SecurityScheme(
-        name = "bearerAuth",
-        type = SecuritySchemeType.HTTP,
-        scheme = "bearer",
-        bearerFormat = "JWT"
-)
-public class SwaggerConfig {
-
-}
